fixing compiler warning
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@5423 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/modules/trackFitChecker/src/TrackFitCheckerModule.cc
@@ -1670,7 +1670,7 @@ int TrackFitCheckerModule::trunctatedMeanAndStd(std::vector<double> data, const 
 {
 
   const int n = data.size();
-  double truncN = -1;
+  int truncN = -1;
 //  if ((n* cutRatio < 2 and symmetric == true) or(n* cutRatio < 1 and symmetric == false)) {
 //    return 0;
 //  }
